Refactor: Message에서 컴포넌트 분리
