+++ src/components/Message/MessageContent.tsx
@@ -12,9 +12,9 @@ const MessageContent = React.memo(
       <article className="border-b border-gray-c8 py-4 px-3">
         <div className="flex items-center justify-between mb-1">
           {isOutgoingMessage ? (
-            <span className="text-highlight">보낸 메시지</span>
+            <span className="text-highlight font-medium">보낸 메시지</span>
           ) : (
-            <span className="text-[#12AA5F]">받은 메세지</span>
+            <span className="text-[#12AA5F] font-medium">받은 메세지</span>
           )}
           <span className="text-sm text-gray-6c dark:text-gray-ee">
             {new Date(date).toLocaleString().slice(0, -3)}
